Elide numeric for loop's step if possible
Antifennel will always include the step parameter to the for loops it
generates, regardless of whether or not the step is necessary.
This creates output that is technically correct, but isn't formatted
in an ideal way. After this patch, antifennel will generate loops like
(for [i 1 10] ...) instead of (for [i 1 10 1] ...) if the step is 1.